ResuMatch/requirements.txt
streamlit
PyPDF2
pdfplumber
python-docx
numpy>=1.26.4
scipy>=1.11.4
scikit-learn>=1.4.2
nltk
langdetect
mistralai
python-dotenv
